requirements.txt
Flask==2.2.5
numpy==1.24.2
opencv_python_headless==4.7.0.72
pyserial==3.5
bcrypt==4.0.1
pyttsx3==2.90
requests==2.28.2
sinricpro==2.7.3